requirements.txt
aiofiles
httpx
Pillow
requests
python-dotenv
opencv-python
git+https://github.com/edenartlab/eden-sdk-py.git